Implement more OOP
